requirements.txt
yt-dlp
openai-whisper
moviepy